Finished the "Build Your First App" tutorial.
Had a few ideas for a project I wanted to undertake this year, and this helped me draw some lines
to getting some more work done on that front, plus a few sketches with a pen and paper helps too.

"Look Ma, I'm an Android developer!" hahaha, baby steps, young padawan, baby steps...

Ran into an issue, as I wasn't intially planning on putting this up on GH, but had to invalidate
AS's cache, clean the project, and then rebuild the APK(s), to get it running on a donated Moto X
(2013) again. SO (https://stackoverflow.com/questions/42219784/installation-failed-with-message-invalid-file)
to the resucue!